Rename metric package to metrics
