+++ pkg/metrics/grafana.go
@@ -1,4 +1,4 @@
-package metric
+package metrics
 
 import (
 	"context"
